--- dfusion_rust_core/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "dfusion_core"
-version = "0.1.0"
-authors = ["Alexander Herrmann <alex@gnosis.io>", "Ben Smith <ben@gnosis.io>", "Felix Leupold <felix@gnosis.io>"]
-edition = "2018"
-
-[dependencies]
-log = "0.4.8"
-mockall = "0.6.0"
-serde_derive = "1.0"
-serde = "1.0"
-web3 = "0.8.0"
