--- dfusion_rust_core/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod models;
