+++ driver/src/contracts/stablex_contract.rs
@@ -3,7 +3,6 @@
 use std::collections::HashMap;
 use std::env;
 
-use dfusion_core::models::{AccountState, Order, Solution};
 use lazy_static::lazy_static;
 #[cfg(test)]
 use mockall::automock;
@@ -13,6 +12,7 @@ use web3::types::{H160, U128, U256};
 use crate::contracts;
 use crate::contracts::stablex_auction_element::StableXAuctionElement;
 use crate::error::DriverError;
+use crate::models::{AccountState, Order, Solution};
 use crate::util::FutureWaitExt;
 
 type Result<T> = std::result::Result<T, DriverError>;
@@ -213,8 +213,8 @@ fn encode_execution_for_contract(
 #[cfg(test)]
 pub mod tests {
     use super::*;
-    use dfusion_core::models::util::map_from_slice;
-    use dfusion_core::models::BatchInformation;
+    use crate::models::BatchInformation;
+    use crate::util::map_from_slice;
 
     #[test]
     #[should_panic]
