+++ driver/src/models/mod.rs
@@ -1,7 +1,6 @@
 pub mod account_state;
 pub mod order;
 pub mod solution;
-pub mod util;
 
 pub use crate::models::account_state::AccountState;
 pub use crate::models::order::BatchInformation;
